Calculate the logsum for the hedonic price. It is a taz level average of the hedonic price that is used. To simplify the computation, a subset of the taz level population is used with adequate weights

+++ dev/Basic/long/database/entity/TazLogsumWeight.cpp
@@ -31,7 +31,7 @@ int TazLogsumWeight::getIndividualId() const
 	return individualId;
 }
 
-int TazLogsumWeight::getWeight() const
+double TazLogsumWeight::getWeight() const
 {
 	return weight;
 }
